requirements.txt
Django==4.2.7
djangorestframework==3.14.0
langchain==0.1.0
langchain-community==0.0.10
sentence-transformers==2.2.2
faiss-cpu>=1.9.0
python-dotenv==1.0.0
Pillow>=10.0.0
transformers>=4.30.0
torch>=2.0.0
numpy>=1.24.0

